pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "tasks*", "evaluator*"]

[project]
name = "ga-bench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.0.0",
    "typer>=0.15.0",
]

[dependency-groups]
langchain = [
    "deepagents>=0.4.1",
    "langchain>=1.2.10",
    "langchain-anthropic>=1.3.3",
    "langchain-google-vertexai>=3.2.2",
    "openevals>=0.0.11",
]
claude = [
    "claude-agent-sdk>=0.1.38",
]
dev = [
    "pre-commit>=4.0.0",
    "ruff>=0.9.0",
    "ty>=0.0.1a3",
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]

[tool.ruff.lint.isort]
known-first-party = ["agents", "evaluator", "tasks"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ty.environment]
python-version = "3.13"
